+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 - Webtorrent tracker for P2P video feature
 - eslint sort the modules name (#2338)
 - Add retention date and s3 lifecycle rules to classroom / video
+- Add a command to delete expired classrooms / videos
 
 ### Changed